This commit is a very early implementation of the metadata to HTML XSL transformations. XSL transformations currently only work for deegree records, as the xsl needs to be modified for different metadata formats.
Also, the XSL is not yet complete for degree - it still needs to be updated for several fields and set up for CSS.

For the HTML results, the changes are currently being populated in the same div that the catalogs are originally displayed in. This is being fixed to pop up in a new tab.